+++ android/app/src/main/kotlin/com/oehamilton/cash_flow_manager/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.oehamilton.cash_flow_manager
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()